feat(拍卖行): 完成竞价、一口价、取消上架功能

+++ client/src/main/java/client/cmd/DurabilityDeficiencyResultClient.java
@@ -18,7 +18,7 @@ public class DurabilityDeficiencyResultClient implements ICmd<GameMsg.Durability
         }
 
         Role role = Role.getInstance();
-        int userEquipmentId = durabilityDeficiencyResult.getUserEquipmentId();
+        long userEquipmentId = durabilityDeficiencyResult.getUserEquipmentId();
         int propsId = durabilityDeficiencyResult.getPropsId();
 
         Props props = GameData.getInstance().getPropsMap().get(propsId);
