+++ client/src/main/java/client/scene/GameData.java
@@ -269,7 +269,7 @@ public class GameData {
                 if (equipmentEntity.getPropsId().equals(value.getId())) {
                     EquipmentType equipmentType = EquipmentType.valueOf(equipmentEntity.getType());
                     value.setPropsProperty(
-                            new Equipment(equipmentEntity.getId(), equipmentEntity.getPropsId(),
+                            new Equipment((long) equipmentEntity.getId(), equipmentEntity.getPropsId(),
                                     equipmentEntity.getDurability(), equipmentEntity.getDamage(), equipmentType));
                     break;
                 }
@@ -278,7 +278,7 @@ public class GameData {
         for (PotionEntity potionEntity : potionEntityList) {
             for (Props value : propsMap.values()) {
                 if (potionEntity.getPropsId().equals(value.getId())) {
-                    value.setPropsProperty(new Potion(potionEntity.getId(), potionEntity.getPropsId(), potionEntity.getCdTime(),
+                    value.setPropsProperty(new Potion((long) potionEntity.getId(), potionEntity.getPropsId(), potionEntity.getCdTime(),
                             potionEntity.getInfo(), potionEntity.getResumeFigure(), potionEntity.getPercent()));
                 }
             }
